+++ packages/polythene-css-classes/menu.js
@@ -10,12 +10,13 @@ export default {
   backdrop:    "pe-menu__backdrop",
 
   // states
-  permanent:   "pe-menu--permanent",
-  floating:    "pe-menu--floating",
-  width_auto:  "pe-menu--width-auto",
-  width_n:     "pe-menu--width-",
-  origin:      "pe-menu--origin",
-  visible:     "pe-menu--visible",
+  floating:     "pe-menu--floating",
+  origin:       "pe-menu--origin",
+  permanent:    "pe-menu--permanent",
+  showBackdrop: "pe-menu--backdrop",
+  visible:      "pe-menu--visible",
+  width_auto:   "pe-menu--width-auto",
+  width_n:      "pe-menu--width-",
 
   // lookup
   listTile:         listTileClasses.component,